+++ packages/@markuplint/rules/src/wai-aria/checkings/value.spec.ts
@@ -1,0 +1,33 @@
+import { checkAriaValue } from './value';
+
+describe('checkAriaValue', () => {
+	test('token', () => {
+		expect(checkAriaValue('token', 'a', ['a'])).toBe(true);
+		expect(checkAriaValue('token', 'a', ['b'])).toBe(false);
+	});
+
+	test('token list', () => {
+		expect(checkAriaValue('token list', 'a', ['a'])).toBe(true);
+		expect(checkAriaValue('token list', 'a', ['b'])).toBe(false);
+		expect(checkAriaValue('token list', 'a b', ['a'])).toBe(false);
+		expect(checkAriaValue('token list', 'a b', ['b'])).toBe(false);
+		expect(checkAriaValue('token list', 'a b', ['a', 'b'])).toBe(true);
+		expect(checkAriaValue('token list', 'a b', ['a', 'b', 'c'])).toBe(true);
+		expect(checkAriaValue('token list', 'a b d', ['a', 'b', 'c'])).toBe(false);
+	});
+
+	test('integer', () => {
+		expect(checkAriaValue('integer', '1', [])).toBe(true);
+		expect(checkAriaValue('integer', '-1', [])).toBe(true);
+		expect(checkAriaValue('integer', '-1.1', [])).toBe(false);
+		expect(checkAriaValue('integer', '-1.1', [])).toBe(false);
+	});
+
+	test('number', () => {
+		expect(checkAriaValue('number', '1', [])).toBe(true);
+		expect(checkAriaValue('number', '-1', [])).toBe(true);
+		expect(checkAriaValue('number', '-1.1', [])).toBe(true);
+		expect(checkAriaValue('number', '-1.1', [])).toBe(true);
+		expect(checkAriaValue('number', '-1.1.1', [])).toBe(false);
+	});
+});
